+++ docs/release/V8/changelog.md
@@ -7,7 +7,10 @@
 链接: [https://caiyun.139.com/m/i?135CeBMHACC6p](https://caiyun.139.com/m/i?135CeBMHACC6p)
 
 
-- 新增[精选应用-自由笔记]，修改系统日期为 2025-2-14 上午 10:30 分即可领取永久的自由笔记免费使用资格。
+- 修复 Android 15 Hyper OS 2 下微信概率竖屏转横屏的问题
+- 修复 Android 15 Hyper OS 2 下安全服务概率竖屏转横屏的问题
+- 修复 Android 15 Hyper OS 2 下小米浏览器概率竖屏转横屏的问题
+- 修复 Android 15 Hyper OS 2 下超级小爱概率竖屏转横屏的问题
 
 
 如果是处于测试版(Beta)的模块，可能存在较多的不稳定性以及异常BUG，如遇问题可反馈，反馈Q群：277757185